refactor: scaffold_messenger_service.dartをディレクトリ移動
